src/agentic-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymavlink==2.4.39
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.4
google-generativeai==0.5.2